Move download button to the left of Bookmark button
When it was added to the left of Share buttons, in Twitter home page,
sometimes was added to the left of some Bookmark buttons, perhaps
because some built-in buttons were added dynamically too.

So instead it's more stable to place download buttons relative to
Bookmark buttons.

Finally since Bookmark buttons in the home page don't have counters,
download buttons are now moved to their left to be grouped with other
buttons that always have counters.